requirements_csv_import.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
